// apps/mobile/lib/gen/model/expenseCategory.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Core API
 * Core API
 * OpenAPI spec version: 0.0.1
 */
import type { ExpenseSubcategory } from './expenseSubcategory';

export interface ExpenseCategory {
  id: string;
  label: string;
  color: string;
  sort_order: number;
  subcategories: ExpenseSubcategory[];
}